perf: more optimizations

+++ contracts/SlotReservations.sol
@@ -17,7 +17,7 @@ abstract contract SlotReservations {
 
   function _slotIsFree(SlotId slotId) internal view virtual returns (bool);
 
-  function reserveSlot(RequestId requestId, uint256 slotIndex) public {
+  function reserveSlot(RequestId requestId, uint64 slotIndex) public {
     require(canReserveSlot(requestId, slotIndex), "Reservation not allowed");
 
     SlotId slotId = Requests.slotId(requestId, slotIndex);
@@ -30,7 +30,7 @@ abstract contract SlotReservations {
 
   function canReserveSlot(
     RequestId requestId,
-    uint256 slotIndex
+    uint64 slotIndex
   ) public view returns (bool) {
     address host = msg.sender;
     SlotId slotId = Requests.slotId(requestId, slotIndex);
@@ -41,5 +41,5 @@ abstract contract SlotReservations {
       (!_reservations[slotId].contains(host));
   }
 
-  event SlotReservationsFull(RequestId indexed requestId, uint256 slotIndex);
+  event SlotReservationsFull(RequestId indexed requestId, uint64 slotIndex);
 }
